+++ src/promptflow/promptflow/_constants.py
@@ -16,3 +16,14 @@ ERROR_RESPONSE_COMPONENT_NAME = "promptflow"
 EXTENSION_UA = "prompt-flow-extension"
 
 DEFAULT_ENCODING = "utf-8"
+
+# Constants related to execution
+LINE_NUMBER_KEY = "line_number"  # Using the same key with portal.
+LINE_TIMEOUT_SEC = 600
+
+
+class FlowLanguage:
+    """The enum of tool source type."""
+
+    Python = "python"
+    Csharp = "csharp"