+++ src/promptflow/promptflow/executor/flow_executor.py
@@ -1,6 +1,7 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
+import asyncio
 import copy
 import functools
 import inspect
@@ -9,10 +10,11 @@ import uuid
 from pathlib import Path
 from threading import current_thread
 from types import GeneratorType
-from typing import AbstractSet, Any, Callable, Dict, List, Mapping, Optional, Tuple
+from typing import Any, Callable, Dict, List, Mapping, Optional, Tuple
 
 import yaml
 
+from promptflow._constants import LINE_NUMBER_KEY, LINE_TIMEOUT_SEC
 from promptflow._core._errors import NotSupported, UnexpectedError
 from promptflow._core.cache_manager import AbstractCacheManager
 from promptflow._core.flow_execution_context import FlowExecutionContext
@@ -20,35 +22,35 @@ from promptflow._core.metric_logger import add_metric_logger, remove_metric_logg
 from promptflow._core.openai_injector import inject_openai_api
 from promptflow._core.operation_context import OperationContext
 from promptflow._core.run_tracker import RunTracker
-from promptflow._core.tool import ToolInvoker, STREAMING_OPTION_PARAMETER_ATTR
+from promptflow._core.tool import STREAMING_OPTION_PARAMETER_ATTR, ToolInvoker
 from promptflow._core.tools_manager import ToolsManager
 from promptflow._utils.context_utils import _change_working_dir
-from promptflow._utils.logger_utils import flow_logger, logger
-from promptflow._utils.multimedia_utils import (
-    load_multimedia_data,
-    load_multimedia_data_recursively
+from promptflow._utils.execution_utils import (
+    apply_default_value_for_input,
+    collect_lines,
+    get_aggregation_inputs_properties,
 )
+from promptflow._utils.logger_utils import flow_logger, logger
+from promptflow._utils.multimedia_utils import load_multimedia_data, load_multimedia_data_recursively
 from promptflow._utils.utils import transpose
 from promptflow.contracts.flow import Flow, FlowInputDefinition, InputAssignment, InputValueType, Node
 from promptflow.contracts.run_info import FlowRunInfo, Status
 from promptflow.contracts.run_mode import RunMode
 from promptflow.exceptions import PromptflowException
 from promptflow.executor import _input_assignment_parser
+from promptflow.executor._async_nodes_scheduler import AsyncNodesScheduler
 from promptflow.executor._errors import NodeOutputNotFound, OutputReferenceNotExist, SingleNodeValidationError
 from promptflow.executor._flow_nodes_scheduler import (
     DEFAULT_CONCURRENCY_BULK,
     DEFAULT_CONCURRENCY_FLOW,
     FlowNodesScheduler,
 )
-from promptflow.executor._result import AggregationResult, BulkResult, LineResult
+from promptflow.executor._result import AggregationResult, LineResult
 from promptflow.executor._tool_invoker import DefaultToolInvoker
 from promptflow.executor._tool_resolver import ToolResolver
 from promptflow.executor.flow_validator import FlowValidator
 from promptflow.storage import AbstractRunStorage
 from promptflow.storage._run_storage import DefaultRunStorage
-
-LINE_NUMBER_KEY = "line_number"  # Using the same key with portal.
-LINE_TIMEOUT_SEC = 600
 
 
 class FlowExecutor:
@@ -122,7 +124,7 @@ class FlowExecutor:
         self._flow = flow
         self._flow_id = flow.id or str(uuid.uuid4())
         self._connections = connections
-        self._aggregation_inputs_references = self._get_aggregation_inputs_properties(flow)
+        self._aggregation_inputs_references = get_aggregation_inputs_properties(flow)
         self._aggregation_nodes = {node.name for node in self._flow.nodes if node.aggregation}
         if worker_count is not None:
             self._worker_count = worker_count
@@ -195,8 +197,32 @@ class FlowExecutor:
         :return: A new instance of FlowExecutor.
         :rtype: ~promptflow.executor.flow_executor.FlowExecutor
         """
-        working_dir = Flow._resolve_working_dir(flow_file, working_dir)
         flow = Flow.from_yaml(flow_file, working_dir=working_dir)
+        return cls._create_from_flow(
+            flow_file=flow_file,
+            flow=flow,
+            connections=connections,
+            working_dir=working_dir,
+            storage=storage,
+            raise_ex=raise_ex,
+            node_override=node_override,
+            line_timeout_sec=line_timeout_sec,
+        )
+
+    @classmethod
+    def _create_from_flow(
+        cls,
+        flow: Flow,
+        connections: dict,
+        working_dir: Optional[Path],
+        *,
+        flow_file: Optional[Path] = None,
+        storage: Optional[AbstractRunStorage] = None,
+        raise_ex: bool = True,
+        node_override: Optional[Dict[str, Dict[str, Any]]] = None,
+        line_timeout_sec: int = LINE_TIMEOUT_SEC,
+    ):
+        working_dir = Flow._resolve_working_dir(flow_file, working_dir)
         if node_override:
             flow = flow._apply_node_overrides(node_override)
         flow = flow._apply_default_node_variants()
@@ -293,9 +319,7 @@ class FlowExecutor:
             )
         # Only load the node's referenced flow inputs
         node_referenced_flow_inputs = FlowExecutor._get_node_referenced_flow_inputs(node, flow.inputs)
-        inputs_with_default_value = FlowExecutor._apply_default_value_for_input(
-            node_referenced_flow_inputs, flow_inputs
-        )
+        inputs_with_default_value = apply_default_value_for_input(node_referenced_flow_inputs, flow_inputs)
         converted_flow_inputs_for_node = FlowValidator.convert_flow_inputs_for_node(
             flow, node, inputs_with_default_value
         )
@@ -329,23 +353,24 @@ class FlowExecutor:
         storage = DefaultRunStorage(base_dir=working_dir, sub_dir=Path(sub_dir))
         run_tracker = RunTracker(storage)
         with run_tracker.node_log_manager:
-            ToolInvoker.activate(DefaultToolInvoker())
-
             # Will generate node run in context
             context = FlowExecutionContext(
                 name=flow.name,
                 run_tracker=run_tracker,
                 cache_manager=AbstractCacheManager.init_from_env(),
             )
-            context.current_node = node
-            context.start()
+
             try:
-                resolved_node.callable(**resolved_inputs)
+                if inspect.iscoroutinefunction(resolved_node.callable):
+                    asyncio.run(
+                        context.invoke_tool_async(resolved_node.node, resolved_node.callable, kwargs=resolved_inputs),
+                    )
+                else:
+                    context.invoke_tool(resolved_node.node, resolved_node.callable, kwargs=resolved_inputs)
             except Exception:
-                if raise_ex:
+                if raise_ex:  # Only raise exception when raise_ex is True
                     raise
-            finally:
-                context.end()
+
             node_runs = run_tracker.collect_node_runs()
             if len(node_runs) != 1:
                 # Should not happen except there is bug in run_tracker or thread control.
@@ -403,46 +428,12 @@ class FlowExecutor:
         """
         return self._aggregation_nodes
 
-    @staticmethod
-    def _get_aggregation_inputs_properties(flow: Flow) -> AbstractSet[str]:
-        normal_node_names = {node.name for node in flow.nodes if flow.is_normal_node(node.name)}
-        properties = set()
-        for node in flow.nodes:
-            if node.name in normal_node_names:
-                continue
-            for value in node.inputs.values():
-                if not value.value_type == InputValueType.NODE_REFERENCE:
-                    continue
-                if value.value in normal_node_names:
-                    properties.add(value.serialize())
-        return properties
-
-    def _collect_lines(self, indexes: List[int], kvs: Mapping[str, List]) -> Mapping[str, List]:
-        """Collect the values from the kvs according to the indexes."""
-        return {k: [v[i] for i in indexes] for k, v in kvs.items()}
-
     def _fill_lines(self, indexes, values, nlines):
         """Fill the values into the result list according to the indexes."""
         result = [None] * nlines
         for idx, value in zip(indexes, values):
             result[idx] = value
         return result
-
-    def _handle_line_failures(self, run_infos: List[FlowRunInfo], raise_on_line_failure: bool = False):
-        failed = [i for i, r in enumerate(run_infos) if r.status == Status.Failed]
-        failed_msg = None
-        if len(failed) > 0:
-            failed_indexes = ",".join([str(i) for i in failed])
-            first_fail_exception = run_infos[failed[0]].error["message"]
-            if raise_on_line_failure:
-                failed_msg = "Flow run failed due to the error: " + first_fail_exception
-                raise Exception(failed_msg)
-
-            failed_msg = (
-                f"{len(failed)}/{len(run_infos)} flow run failed, indexes: [{failed_indexes}],"
-                f" exception of index {failed[0]}: {first_fail_exception}"
-            )
-            logger.error(failed_msg)
 
     def _exec_batch_with_process_pool(
         self, batch_inputs: List[dict], run_id, output_dir: Path, validate_inputs: bool = True, variant_id: str = ""
@@ -501,7 +492,7 @@ class FlowExecutor:
             [result.aggregation_inputs for result in results],
             keys=self._aggregation_inputs_references,
         )
-        succeeded_aggregation_inputs = self._collect_lines(succeeded, aggregation_inputs)
+        succeeded_aggregation_inputs = collect_lines(succeeded, aggregation_inputs)
         try:
             aggr_results = self._exec_aggregation(succeeded_inputs, succeeded_aggregation_inputs, run_id)
             logger.info("Finish executing aggregation nodes.")
@@ -647,7 +638,7 @@ class FlowExecutor:
         :rtype: dict
         """
         self._node_concurrency = node_concurrency
-        inputs = FlowExecutor._apply_default_value_for_input(self._flow.inputs, inputs)
+        inputs = apply_default_value_for_input(self._flow.inputs, inputs)
         result = self._exec(inputs)
         #  TODO: remove this line once serving directly calling self.exec_line
         self._add_line_results([result])
@@ -705,10 +696,10 @@ class FlowExecutor:
         :rtype: ~promptflow.executor._result.LineResult
         """
         self._node_concurrency = node_concurrency
-        inputs = FlowExecutor._apply_default_value_for_input(self._flow.inputs, inputs)
+        inputs = apply_default_value_for_input(self._flow.inputs, inputs)
         # For flow run, validate inputs as default
         with self._run_tracker.node_log_manager:
-            # exec_line interface may be called by exec_bulk, so we only set run_mode as flow run when
+            # exec_line interface may be called when executing a batch run, so we only set run_mode as flow run when
             # it is not set.
             operation_context = OperationContext.get_instance()
             operation_context.run_mode = operation_context.get("run_mode", None) or RunMode.Test.name
@@ -734,67 +725,6 @@ class FlowExecutor:
                 for node_run_info in result.node_run_infos.values()
             }
         )
-
-    def exec_bulk(
-        self,
-        inputs: List[Dict[str, Any]],
-        run_id: str = None,
-        validate_inputs: bool = True,
-        raise_on_line_failure: bool = False,
-        node_concurrency=DEFAULT_CONCURRENCY_BULK,
-        output_dir: Path = None,
-    ) -> BulkResult:
-        """The entry points for bulk run execution
-
-        :param inputs: A list of dictionaries containing input data.
-        :type inputs: List[Dict[str, Any]]
-        :param run_id: Run ID.
-        :type run_id: Optional[str]
-        :param validate_inputs: Whether to validate the inputs. Defaults to True.
-        :type validate_inputs: Optional[bool]
-        :param raise_on_line_failure: Whether to raise an exception on line failure. Defaults to False. \
-        [To be deprecated]
-        :type raise_on_line_failure: Optional[bool]
-        :param node_concurrency: The node concurrency. Defaults to DEFAULT_CONCURRENCY_BULK.
-        :type node_concurrency: Optional[int]
-        :return: The bulk result.
-        :rtype: ~promptflow.executor.flow_executor.BulkResult
-        """
-
-        self._node_concurrency = node_concurrency
-        # Apply default value in early stage, so we can use it both in line execution and aggregation nodes execution.
-        inputs = [
-            FlowExecutor._apply_default_value_for_input(self._flow.inputs, each_line_input)
-            for each_line_input in inputs
-        ]
-        run_id = run_id or str(uuid.uuid4())
-        with self._run_tracker.node_log_manager:
-            OperationContext.get_instance().run_mode = RunMode.Batch.name
-            line_results = self._exec_batch_with_process_pool(
-                inputs, run_id, output_dir, validate_inputs=validate_inputs
-            )
-            self._add_line_results(line_results)  # For bulk run, currently we need to add line results to run_tracker
-            self._handle_line_failures([r.run_info for r in line_results], raise_on_line_failure)
-            aggr_results = self._exec_aggregation_with_bulk_results(inputs, line_results, run_id)
-        outputs = [
-            {LINE_NUMBER_KEY: r.run_info.index, **r.output}
-            for r in line_results
-            if r.run_info.status == Status.Completed
-        ]
-        return BulkResult(
-            outputs=outputs,
-            metrics=aggr_results.metrics,
-            line_results=line_results,
-            aggr_results=aggr_results,
-        )
-
-    @staticmethod
-    def _apply_default_value_for_input(inputs: Dict[str, FlowInputDefinition], line_inputs: Mapping) -> Dict[str, Any]:
-        updated_inputs = dict(line_inputs or {})
-        for key, value in inputs.items():
-            if key not in updated_inputs and (value and value.default):
-                updated_inputs[key] = value.default
-        return updated_inputs
 
     @staticmethod
     def _get_node_referenced_flow_inputs(
@@ -943,7 +873,15 @@ class FlowExecutor:
     def _traverse_nodes(self, inputs, context: FlowExecutionContext) -> Tuple[dict, dict]:
         batch_nodes = [node for node in self._flow.nodes if not node.aggregation]
         outputs = {}
-        nodes_outputs, bypassed_nodes = self._submit_to_scheduler(context, inputs, batch_nodes)
+        #  TODO: Use a mixed scheduler to support both async and thread pool mode.
+        should_use_async = all(inspect.iscoroutinefunction(f) for f in self._tools_manager._tools.values())
+        if should_use_async:
+            flow_logger.info("Start executing nodes in async mode.")
+            scheduler = AsyncNodesScheduler(self._tools_manager, self._node_concurrency)
+            nodes_outputs, bypassed_nodes = asyncio.run(scheduler.execute(batch_nodes, inputs, context))
+        else:
+            flow_logger.info("Start executing nodes in thread pool mode.")
+            nodes_outputs, bypassed_nodes = self._submit_to_scheduler(context, inputs, batch_nodes)
         outputs = self._extract_outputs(nodes_outputs, bypassed_nodes, inputs)
         return outputs, nodes_outputs