+++ src/promptflow/tests/sdk_cli_test/e2etests/test_flow_as_func.py
@@ -1,7 +1,9 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
+import shutil
 from pathlib import Path
+from tempfile import TemporaryDirectory
 from types import GeneratorType
 
 import pytest
@@ -9,10 +11,10 @@ import pytest
 from promptflow import load_flow
 from promptflow._sdk._errors import ConnectionNotFoundError, InvalidFlowError
 from promptflow._sdk.entities import CustomConnection
+from promptflow._sdk.operations._flow_context_resolver import FlowContextResolver
+from promptflow._utils.flow_utils import dump_flow_dag, load_flow_dag
 from promptflow.entities import FlowContext
 from promptflow.exceptions import UserErrorException
-
-from ..recording_utilities import RecordStorage
 
 FLOWS_DIR = "./tests/test_configs/flows"
 RUNS_DIR = "./tests/test_configs/runs"
@@ -20,7 +22,7 @@ DATAS_DIR = "./tests/test_configs/datas"
 
 
 @pytest.mark.usefixtures(
-    "use_secrets_config_file", "setup_local_connection", "install_custom_tool_pkg", "recording_injection"
+    "use_secrets_config_file", "recording_injection", "setup_local_connection", "install_custom_tool_pkg"
 )
 @pytest.mark.sdk_test
 @pytest.mark.e2etest
@@ -29,6 +31,7 @@ class TestFlowAsFunc:
         f = load_flow(f"{FLOWS_DIR}/print_env_var")
         result = f(key="unknown")
         assert result["output"] is None
+        assert "line_number" not in result
 
     def test_flow_as_a_func_with_connection_overwrite(self):
         from promptflow._sdk._errors import ConnectionNotFoundError
@@ -40,7 +43,6 @@ class TestFlowAsFunc:
             f(url="https://www.youtube.com/watch?v=o5ZQyXaAv1g")
         assert "Connection 'not_exist' is not found" in str(e.value)
 
-    @pytest.mark.skipif(RecordStorage.is_replaying_mode(), reason="TODO: support customized python tool in future")
     def test_flow_as_a_func_with_connection_obj(self):
         f = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
         f.context.connections = {"hello_node": {"connection": CustomConnection(secrets={"k": "v"})}}
@@ -77,12 +79,11 @@ class TestFlowAsFunc:
             f(text="hello")
         assert "Failed to load python module " in str(e.value)
 
-        f = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
+        f = load_flow(f"{FLOWS_DIR}/print_env_var")
         with pytest.raises(UserErrorException) as e:
             f()
-        assert "Required input(s) ['text'] are missing" in str(e.value)
+        assert "Required input(s) ['key'] are missing" in str(e.value)
 
-    @pytest.mark.skipif(RecordStorage.is_replaying_mode(), reason="Stream not supported in replaying mode.")
     def test_stream_output(self):
         f = load_flow(f"{FLOWS_DIR}/chat_flow_with_python_node_streaming_output")
         f.context.streaming = True
@@ -126,21 +127,115 @@ class TestFlowAsFunc:
         assert "please make sure connection has decrypted secrets to use in flow execution." in str(e)
 
     def test_local_connection_object(self, pf, azure_open_ai_connection):
-        f = load_flow(f"{FLOWS_DIR}/web_classification")
-        f.context.connections = {"classify_with_llm": {"connection": azure_open_ai_connection}}
-        f()
-
+        f = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
         # local connection without secret will lead to error
         connection = pf.connections.get("azure_open_ai_connection", with_secrets=False)
-        f.context.connections = {"classify_with_llm": {"connection": connection}}
+        f.context.connections = {"hello_node": {"connection": connection}}
         with pytest.raises(UserErrorException) as e:
-            f()
+            f(text="hello")
         assert "please make sure connection has decrypted secrets to use in flow execution." in str(e)
 
-    @pytest.mark.skipif(RecordStorage.is_replaying_mode(), reason="Returning dict is not supported for now.")
     def test_non_secret_connection(self):
         f = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
         # execute connection without secrets won't get error since the connection doesn't have scrubbed secrets
         # we only raise error when there are scrubbed secrets in connection
         f.context.connections = {"hello_node": {"connection": CustomConnection(secrets={})}}
         f(text="hello")
+
+    def test_flow_context_cache(self):
+        # same flow context has same hash
+        assert hash(FlowContext()) == hash(FlowContext())
+        # getting executor for same flow will hit cache
+        flow1 = load_flow(f"{FLOWS_DIR}/print_env_var")
+        flow2 = load_flow(f"{FLOWS_DIR}/print_env_var")
+        flow_executor1 = FlowContextResolver.create(
+            flow=flow1,
+        )
+        flow_executor2 = FlowContextResolver.create(
+            flow=flow2,
+        )
+        assert flow_executor1 is flow_executor2
+
+        # getting executor for same flow + context will hit cache
+        flow1 = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
+        flow1.context = FlowContext(connections={"hello_node": {"connection": CustomConnection(secrets={"k": "v"})}})
+        flow2 = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
+        flow2.context = FlowContext(connections={"hello_node": {"connection": CustomConnection(secrets={"k": "v"})}})
+        flow_executor1 = FlowContextResolver.create(
+            flow=flow1,
+        )
+        flow_executor2 = FlowContextResolver.create(
+            flow=flow2,
+        )
+        assert flow_executor1 is flow_executor2
+
+        flow1 = load_flow(f"{FLOWS_DIR}/flow_with_dict_input_with_variant")
+        flow1.context = FlowContext(
+            variant="${print_val.variant1}",
+            connections={"print_val": {"conn": CustomConnection(secrets={"k": "v"})}},
+            overrides={"nodes.print_val.inputs.key": "a"},
+        )
+        flow2 = load_flow(f"{FLOWS_DIR}/flow_with_dict_input_with_variant")
+        flow2.context = FlowContext(
+            variant="${print_val.variant1}",
+            connections={"print_val": {"conn": CustomConnection(secrets={"k": "v"})}},
+            overrides={"nodes.print_val.inputs.key": "a"},
+        )
+        flow_executor1 = FlowContextResolver.create(flow=flow1)
+        flow_executor2 = FlowContextResolver.create(flow=flow2)
+        assert flow_executor1 is flow_executor2
+
+    def test_flow_cache_not_hit(self):
+        with TemporaryDirectory() as tmp_dir:
+            shutil.copytree(f"{FLOWS_DIR}/print_env_var", f"{tmp_dir}/print_env_var")
+            flow_path = Path(f"{tmp_dir}/print_env_var")
+            # load same file with different content will not hit cache
+            flow1 = load_flow(flow_path)
+            # update content
+            _, flow_dag = load_flow_dag(flow_path)
+            flow_dag["inputs"] = {"key": {"type": "string", "default": "key1"}}
+            dump_flow_dag(flow_dag, flow_path)
+            flow2 = load_flow(f"{tmp_dir}/print_env_var")
+            flow_executor1 = FlowContextResolver.create(
+                flow=flow1,
+            )
+            flow_executor2 = FlowContextResolver.create(
+                flow=flow2,
+            )
+            assert flow_executor1 is not flow_executor2
+
+    def test_flow_context_cache_not_hit(self):
+        flow1 = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
+        flow1.context = FlowContext(connections={"hello_node": {"connection": CustomConnection(secrets={"k": "v"})}})
+        flow2 = load_flow(f"{FLOWS_DIR}/flow_with_custom_connection")
+        flow2.context = FlowContext(connections={"hello_node": {"connection": CustomConnection(secrets={"k2": "v"})}})
+        flow_executor1 = FlowContextResolver.create(
+            flow=flow1,
+        )
+        flow_executor2 = FlowContextResolver.create(
+            flow=flow2,
+        )
+        assert flow_executor1 is not flow_executor2
+
+        flow1 = load_flow(f"{FLOWS_DIR}/flow_with_dict_input_with_variant")
+        flow1.context = FlowContext(
+            variant="${print_val.variant1}",
+            connections={"print_val": {"conn": CustomConnection(secrets={"k": "v"})}},
+            overrides={"nodes.print_val.inputs.key": "a"},
+        )
+        flow2 = load_flow(f"{FLOWS_DIR}/flow_with_dict_input_with_variant")
+        flow2.context = FlowContext(
+            variant="${print_val.variant1}",
+            connections={"print_val": {"conn": CustomConnection(secrets={"k": "v"})}},
+            overrides={"nodes.print_val.inputs.key": "b"},
+        )
+        flow_executor1 = FlowContextResolver.create(flow=flow1)
+        flow_executor2 = FlowContextResolver.create(flow=flow2)
+        assert flow_executor1 is not flow_executor2
+
+    @pytest.mark.timeout(10)
+    def test_flow_as_func_perf_test(self):
+        # this test should not take long due to caching logic
+        f = load_flow(f"{FLOWS_DIR}/print_env_var")
+        for i in range(100):
+            f(key="key")