+++ src/promptflow/CHANGELOG.md
@@ -1,14 +1,21 @@
 # Release History
 
 ## 1.1.0 (Upcoming)
+
 ### Features Added
-- Add `pfazure flow list` to list flows from Azure AI.
+- Add `pfazure flow show/list` to show or list flows from Azure AI.
+- Display node status in run visualize page graph view.
+- Add support for image input and output in prompt flow.
+- [SDK/CLI] SDK/CLI will collect telemetry by default, user can use `pf config set telemetry.enabled=false` to opt out.
 
 ### Bugs Fixed
+
+- [SDK/CLI] Fix node test with dict node input will raise "Required input(s) missing".
 
 ### Improvements
 - Force 'az login' if using azureml connection provider in cli command.
 - Add env variable 'PF_NO_INTERACTIVE_LOGIN' to disable interactive login if using azureml connection provider in promptflow sdk.
+- Improved CLI invoke time.
 
 ## 1.0.0 (2023.11.09)